upgraded from react-router-dom v6.2 to v6.9 which provides new api
