+++ RootLayout.jsx
@@ -1,0 +1,37 @@
+import React, { useEffect } from "react";
+import { Outlet } from "react-router-dom";
+import { AnimatePresence } from "framer-motion";
+import { Header } from "./components";
+import { useStateValue } from "./context/StateProvider";
+import { getAllFoodItems } from "./utils/firebaseFunctions";
+import { actionType } from "./context/reducer";
+
+const RootLayout = () => {
+  const [{ foodItems }, dispatch] = useStateValue();
+
+  const fetchData = async () => {
+    await getAllFoodItems().then((data) => {
+      dispatch({
+        type: actionType.SET_FOOD_ITEMS,
+        foodItems: data,
+      });
+    });
+  };
+
+  useEffect(() => {
+    fetchData();
+  }, []);
+  return (
+    <AnimatePresence exitBeforeEnter>
+      <div className="w-screen h-auto flex flex-col bg-primary">
+        <Header />
+
+        <main className="mt-14 md:mt-20 px-4 md:px-16 py-4 w-full">
+          <Outlet />
+        </main>
+      </div>
+    </AnimatePresence>
+  );
+};
+
+export default RootLayout;
