+++ package.json
@@ -11,7 +11,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
-    "react-router-dom": "^6.2.2",
+    "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.0",
     "tailwind-scrollbar": "^1.3.1",
     "web-vitals": "^2.1.0"
